refactor: Move fake implementations to testFixtures
This commit refactors the project structure by moving fake implementations
of repositories and utilities from the `domain:testing` module to the
`testFixtures` source set of their respective domain modules.

This change centralizes test utility code closer to the modules that use
them, improving modularity and maintainability. The `domain:testing`
module has been removed as it is no longer needed.

Key changes:

- **Modules:** The `domain:testing` module is removed.
- **`build.gradle.kts` files:**
    - `domain/common`, `domain/framework`, `domain/repository` now apply the
      `java-test-fixtures` plugin.
    - `domain/use-case/build.gradle.kts` now depends on `testFixtures` of
      `domain.common`, `domain.framework`, and `domain.repository`.
- **File Renames and Moves:**
    - `domain/testing/src/main/kotlin/com/eblan/launcher/domain/testing/common/FakeIconKeyGenerator.kt`
      is moved and renamed to
      `domain/common/src/testFixtures/kotlin/com/eblan/launcher/domain/common/FakeIconKeyGenerator.kt`.
    - `domain/testing/src/main/kotlin/com/eblan/launcher/domain/testing/framework/FakeFileManager.kt`
      is moved and renamed to
      `domain/framework/src/testFixtures/kotlin/com/eblan/launcher/domain/framework/FakeFileManager.kt`.
    - `domain/testing/src/main/kotlin/com/eblan/launcher/domain/testing/repository/FakeGridRepository.kt`
      is moved and renamed to
      `domain/repository/src/testFixtures/kotlin/com/eblan/launcher/domain/repository/FakeGridRepository.kt`.
    - `domain/testing/src/main/kotlin/com/eblan/launcher/domain/testing/repository/FakeUserDataRepository.kt`
      is moved and renamed to
      `domain/repository/src/testFixtures/kotlin/com/eblan/launcher/domain/repository/FakeUserDataRepository.kt`.
- **`settings.gradle.kts`:** The `domain:testing` module is removed from
  `settings.gradle.kts`.
- **`.idea/gradle.xml`:** The `domain:testing` module is removed from the
  Gradle module list.
- **Test File Updates:** Test files in `domain/use-case/src/test` are updated
  to reflect the new locations of fake implementations.
