+++ domain/repository/src/testFixtures/kotlin/com/eblan/launcher/domain/repository/FakeGridRepository.kt
@@ -15,7 +15,7 @@
  *   limitations under the License.
  *
  */
-package com.eblan.launcher.domain.testing.repository
+package com.eblan.launcher.domain.repository
 
 import com.eblan.launcher.domain.model.ApplicationInfoGridItem
 import com.eblan.launcher.domain.model.FolderGridItem
@@ -25,7 +25,6 @@ import com.eblan.launcher.domain.model.GridItems
 import com.eblan.launcher.domain.model.ShortcutConfigGridItem
 import com.eblan.launcher.domain.model.ShortcutInfoGridItem
 import com.eblan.launcher.domain.model.WidgetGridItem
-import com.eblan.launcher.domain.repository.GridRepository
 import kotlinx.coroutines.flow.Flow
 import kotlinx.coroutines.flow.MutableStateFlow
 import kotlinx.coroutines.flow.update
