+++ settings.gradle.kts
@@ -35,7 +35,6 @@ include(":domain:framework")
 include(":domain:grid")
 include(":domain:model")
 include(":domain:repository")
-include(":domain:testing")
 include(":domain:use-case")
 include(":feature:action")
 include(":feature:edit-application-info")
